chore(lint): fix 61 clippy errors — type aliases, FromStr, struct params, mechanical fixes
- 31x type_complexity: type aliases for sqlx query_as tuples
- 8x sort_by_key: replace sort_by with idiomatic sort_by_key + Reverse
- 5x should_implement_trait: proper impl FromStr on Currency, Platform, ProviderName, Verdict, SkillDomain
- 5x too_many_arguments: refactor to struct params (ChallengeSubmissionInput, SendWithLogParams, NotificationPayload, StripeSubscriptionUpsert, LedgerHashInput)
- 6x doc_list_item_overindented + 2x doc_list_item_without_indentation
- 3x inconsistent_digit_grouping
- 1x manual_clamp, 1x if_same_then_else, 1x field_reassign_with_default
- 7x await_holding_lock in tests: std::sync::Mutex -> tokio::sync::Mutex
- 1x #[allow] on grpc mod for auto-generated code (justified)

No behavioural change. cargo clippy --tests -- -D warnings: 0 errors.

+++ src/routes/bounties.rs
@@ -318,7 +318,7 @@ async fn create_bounty(
     let enterprise_id = current_enterprise_for(&state.db, auth.user_id).await?;
     let reward = BigDecimal::from_str(&body.reward_credits)
         .map_err(|_| AppError::Validation("invalid reward_credits amount".into()))?;
-    if reward <= BigDecimal::from(0) {
+    if reward <= 0 {
         return Err(AppError::Validation("reward_credits must be > 0".into()));
     }
 
@@ -883,7 +883,7 @@ async fn handle_pull_request_event(state: &AppState, payload: &Value) -> Result<
         // pas le reward brut. Ex bounty 500€ → wallet 460€ EUR.
         let rate_bd = BigDecimal::try_from(rate).unwrap_or(BigDecimal::from(0));
         let fiat_amount = &talent_share_bd * &rate_bd;
-        if fiat_amount > BigDecimal::from(0) {
+        if fiat_amount > 0 {
             let entry = crate::services::talent_wallet::LedgerEntry {
                 user_id: talent_user_id,
                 delta: &fiat_amount,
